решение задачи 2

+++ Семинар 1/Задача 2/Program.cs
@@ -1,8 +1,8 @@
 ﻿Console.WriteLine("Введите число a");
-int a = Convert.ToInt32(Console.ReadLine());
+double a = Convert.ToInt32(Console.ReadLine());
 
 Console.WriteLine("Введите число b");
-int b = Convert.ToInt32(Console.ReadLine());
+double b = Convert.ToInt32(Console.ReadLine());
 
 if (a*a == b)
 {
